Move the MIDI queue outside of VST
